Extracted common code to an abstract class
