+++ tradingcards-plugin/src/main/java/net/tinetwork/tradingcards/tradingcardsplugin/commands/migrate/CardMigratorBukkitRunnable.java
@@ -1,0 +1,13 @@
+package net.tinetwork.tradingcards.tradingcardsplugin.commands.migrate;
+
+import org.bukkit.scheduler.BukkitRunnable;
+
+/**
+ * @author sarhatabaot
+ */
+public class CardMigratorBukkitRunnable extends BukkitRunnable {
+    @Override
+    public void run() {
+
+    }
+}
